requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
cssselect==1.1.0
dateparser==1.0.0
fake-headers==1.0.2
fake-useragent==0.1.11
html5lib==1.1
idna==2.10
lxml==4.6.1
multiprocessing-logging==0.3.1
parse==1.18.0
psycopg2-binary==2.8.6
pyee==7.0.4
pyppeteer==0.2.2
pyquery==1.4.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
regex==2020.11.13
requests==2.24.0
requests-html==0.10.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tqdm==4.52.0
tzlocal==2.1
urllib3==1.25.11
w3lib==1.22.0
webencodings==0.5.1
websockets==8.1
